# web/openapi.json
{
  "components": {
    "schemas": {
...
      "CommonResponse_Any_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "title": "Data"
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse[Any]",
        "type": "object"
      },
...
      "CommonResponse_DeleteWorkProjectResponse_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DeleteWorkProjectResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse[DeleteWorkProjectResponse]",
        "type": "object"
      },
...
      "CommonResponse_QueryWorkProjectAssetsResponse_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/QueryWorkProjectAssetsResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse[QueryWorkProjectAssetsResponse]",
        "type": "object"
      },
      "CommonResponse_QueryWorkProjectFindingsResponse_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/QueryWorkProjectFindingsResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse[QueryWorkProjectFindingsResponse]",
        "type": "object"
      },
      "CommonResponse_QueryWorkProjectsResponse_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/QueryWorkProjectsResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse[QueryWorkProjectsResponse]",
        "type": "object"
      },
...
      "CommonResponse_WorkProjectGraphSnapshotSchema_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkProjectGraphSnapshotSchema"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse[WorkProjectGraphSnapshotSchema]",
        "type": "object"
      },
      "CommonResponse_WorkProjectSchema_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkProjectSchema"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse[WorkProjectSchema]",
        "type": "object"
      },
...
      "CreateWorkProjectRequest": {
        "properties": {
          "assets": {
            "items": {
              "$ref": "#/components/schemas/WorkProjectAssetRequest"
            },
            "maxItems": 500,
            "minItems": 1,
            "title": "Assets",
            "type": "array"
          },
          "description": {
            "default": "",
            "maxLength": 2000,
            "title": "Description",
            "type": "string"
          },
          "name": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "owner_user_ids": {
            "items": {
              "type": "integer"
            },
            "maxItems": 100,
            "title": "Owner User Ids",
            "type": "array"
          },
          "sandbox_container_id": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sandbox Container Id"
          },
          "type": {
            "$ref": "#/components/schemas/WorkProjectType",
            "default": "penetration_test"
          }
        },
        "required": [
          "name",
          "assets"
        ],
        "title": "CreateWorkProjectRequest",
        "type": "object"
      },
...
      "DeleteWorkProjectResponse": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          }
        },
        "required": [
          "id"
        ],
        "title": "DeleteWorkProjectResponse",
        "type": "object"
      },
...
      "QueryWorkProjectAssetsResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/WorkProjectAssetSchema"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "page",
          "size",
          "total",
          "items"
        ],
        "title": "QueryWorkProjectAssetsResponse",
        "type": "object"
      },
      "QueryWorkProjectFindingsResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/WorkProjectFindingSchema"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "page",
          "size",
          "total",
          "items"
        ],
        "title": "QueryWorkProjectFindingsResponse",
        "type": "object"
      },
      "QueryWorkProjectsResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/WorkProjectSchema"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "page",
          "size",
          "total",
          "items"
        ],
        "title": "QueryWorkProjectsResponse",
        "type": "object"
      },
...
      "SystemUserRole": {
        "enum": [
          "admin",
          "user"
        ],
        "title": "SystemUserRole",
        "type": "string"
      },
...
      "WorkProjectAgentSummarySchema": {
        "properties": {
          "agent_code": {
            "maxLength": 32,
            "minLength": 1,
            "title": "Agent Code",
            "type": "string"
          },
          "summary": {
            "$ref": "#/components/schemas/WorkProjectAgentSummaryContentSchema"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "agent_code"
        ],
        "title": "WorkProjectAgentSummarySchema",
        "type": "object"
      },
      "WorkProjectAssetExtraSchema": {
        "additionalProperties": false,
        "description": "Recon metadata attached to an asset. Kept small on purpose.",
        "properties": {
          "banner": {
            "default": "",
            "maxLength": 512,
            "title": "Banner",
            "type": "string"
          }
        },
        "title": "WorkProjectAssetExtraSchema",
        "type": "object"
      },
      "WorkProjectAssetOrigin": {
        "enum": [
          "scope",
          "discovered"
        ],
        "title": "WorkProjectAssetOrigin",
        "type": "string"
      },
      "WorkProjectAssetRequest": {
        "additionalProperties": false,
        "properties": {
          "extra": {
            "$ref": "#/components/schemas/WorkProjectAssetExtraSchema"
          },
          "host": {
            "default": "",
            "maxLength": 255,
            "title": "Host",
            "type": "string"
          },
          "path": {
            "default": "",
            "maxLength": 500,
            "title": "Path",
            "type": "string"
          },
          "port": {
            "anyOf": [
              {
                "maximum": 65535.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Port"
          },
          "type": {
            "$ref": "#/components/schemas/WorkProjectAssetType"
          }
        },
        "required": [
          "type"
        ],
        "title": "WorkProjectAssetRequest",
        "type": "object"
      },
      "WorkProjectAssetSchema": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by_agent_code": {
            "default": "",
            "title": "Created By Agent Code",
            "type": "string"
          },
          "created_from_session_id": {
            "default": "",
            "title": "Created From Session Id",
            "type": "string"
          },
          "extra": {
            "$ref": "#/components/schemas/WorkProjectAssetExtraSchema"
          },
          "host": {
            "default": "",
            "title": "Host",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "identifier": {
            "title": "Identifier",
            "type": "string"
          },
          "origin": {
            "$ref": "#/components/schemas/WorkProjectAssetOrigin"
          },
          "path": {
            "default": "",
            "title": "Path",
            "type": "string"
          },
          "port": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Port"
          },
          "project_id": {
            "title": "Project Id",
            "type": "integer"
          },
          "type": {
            "$ref": "#/components/schemas/WorkProjectAssetType"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "project_id",
          "type",
          "origin",
          "identifier",
          "created_at",
          "updated_at"
        ],
        "title": "WorkProjectAssetSchema",
        "type": "object"
      },
      "WorkProjectAssetType": {
        "enum": [
          "service",
          "domain",
          "network",
          "binary"
        ],
        "title": "WorkProjectAssetType",
        "type": "string"
      },
...
      "WorkProjectGraphSnapshotSchema": {
        "properties": {
          "attack_path_steps": {
            "items": {
              "$ref": "#/components/schemas/WorkProjectAttackPathStepSchema"
            },
            "title": "Attack Path Steps",
            "type": "array"
          },
          "attack_paths": {
            "items": {
              "$ref": "#/components/schemas/WorkProjectAttackPathSchema"
            },
            "title": "Attack Paths",
            "type": "array"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/WorkProjectGraphEdgeSchema"
            },
            "title": "Edges",
            "type": "array"
          }
        },
        "required": [
          "edges",
          "attack_paths",
          "attack_path_steps"
        ],
        "title": "WorkProjectGraphSnapshotSchema",
        "type": "object"
      },
      "WorkProjectOwnerSchema": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "role": {
            "$ref": "#/components/schemas/SystemUserRole"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "id",
          "role",
          "username"
        ],
        "title": "WorkProjectOwnerSchema",
        "type": "object"
      },
      "WorkProjectSchema": {
        "properties": {
          "agent_summaries": {
            "items": {
              "$ref": "#/components/schemas/WorkProjectAgentSummarySchema"
            },
            "title": "Agent Summaries",
            "type": "array"
          },
          "assets": {
            "items": {
              "$ref": "#/components/schemas/WorkProjectAssetSchema"
            },
            "title": "Assets",
            "type": "array"
          },
          "can_cancel": {
            "default": false,
            "title": "Can Cancel",
            "type": "boolean"
          },
          "can_create_session": {
            "default": false,
            "title": "Can Create Session",
            "type": "boolean"
          },
          "can_retry": {
            "default": false,
            "title": "Can Retry",
            "type": "boolean"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "owner_user_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Owner User Ids",
            "type": "array"
          },
          "owners": {
            "items": {
              "$ref": "#/components/schemas/WorkProjectOwnerSchema"
            },
            "title": "Owners",
            "type": "array"
          },
          "progress": {
            "default": 0,
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Progress",
            "type": "number"
          },
          "sandbox_container_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sandbox Container Id"
          },
          "session_count": {
            "default": 0,
            "title": "Session Count",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/WorkProjectStatus"
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/WorkProjectTaskSchema"
            },
            "title": "Tasks",
            "type": "array"
          },
          "type": {
            "$ref": "#/components/schemas/WorkProjectType"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "owner_user_ids",
          "owners",
          "assets",
          "tasks",
          "agent_summaries",
          "status",
          "type",
          "created_at",
          "updated_at"
        ],
        "title": "WorkProjectSchema",
        "type": "object"
      },
      "WorkProjectStatus": {
        "enum": [
          "working",
          "completed",
          "canceled"
        ],
        "title": "WorkProjectStatus",
        "type": "string"
      },
...
      "WorkProjectType": {
        "enum": [
          "penetration_test",
          "source_code_audit"
        ],
        "title": "WorkProjectType",
        "type": "string"
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Z3r0 - A controlled multi-agent workbench for authorized security assessment, code auditing, internal review, and controlled research.",
    "version": "0.1.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/work-projects": {
      "get": {
        "operationId": "query_work_projects_route_api_work_projects_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 100,
              "minimum": 1,
              "title": "Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "keyword",
            "required": false,
            "schema": {
              "default": "",
              "title": "Keyword",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_QueryWorkProjectsResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Query Work Projects Route",
        "tags": [
          "work-projects"
        ]
      },
      "post": {
        "operationId": "create_work_project_route_api_work_projects_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_WorkProjectSchema_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Create Work Project Route",
        "tags": [
          "work-projects"
        ]
      }
    },
    "/api/work-projects/{id}": {
      "delete": {
        "operationId": "delete_work_project_handler_api_work_projects__id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_DeleteWorkProjectResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Work project not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Delete Work Project Handler",
        "tags": [
          "work-projects"
        ]
      },
      "get": {
        "operationId": "get_work_project_route_api_work_projects__id__get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_WorkProjectSchema_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Work project not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Get Work Project Route",
        "tags": [
          "work-projects"
        ]
      }
    },
...
    "/api/work-projects/{project_id}/assets": {
      "get": {
        "operationId": "query_work_project_assets_route_api_work_projects__project_id__assets_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 100,
              "minimum": 1,
              "title": "Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "keyword",
            "required": false,
            "schema": {
              "default": "",
              "title": "Keyword",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_QueryWorkProjectAssetsResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Query Work Project Assets Route",
        "tags": [
          "work-project-assets"
        ]
      }
    },
    "/api/work-projects/{project_id}/findings": {
      "get": {
        "operationId": "query_work_project_findings_route_api_work_projects__project_id__findings_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 100,
              "minimum": 1,
              "title": "Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "keyword",
            "required": false,
            "schema": {
              "default": "",
              "title": "Keyword",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_QueryWorkProjectFindingsResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Query Work Project Findings Route",
        "tags": [
          "work-project-findings"
        ]
      }
    },
    "/api/work-projects/{project_id}/graph": {
      "get": {
        "operationId": "get_work_project_graph_snapshot_route_api_work_projects__project_id__graph_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_WorkProjectGraphSnapshotSchema_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Get Work Project Graph Snapshot Route",
        "tags": [
          "work-project-graph"
        ]
      }
    }
  }
}